only clear suite collections

+++ fun_test/scripts/tasks/maintenance.py
@@ -181,9 +181,10 @@ class RemoveOldCollections(FunTestCase):
         if collection_names > CheckMongoCollectionCount.MAX_COLLECTIONS:
             collections_to_be_removed = collection_names[0: len(collection_names) - CheckMongoCollectionCount.MAX_COLLECTIONS]
             for collection_to_be_removed in collections_to_be_removed:
-                collection = mongo.get_collection(collection_name=collection_to_be_removed)
-                if collection:
-                    collection.drop()
+                if "s_" in collections_to_be_removed:
+                    collection = mongo.get_collection(collection_name=collection_to_be_removed)
+                    if collection:
+                        collection.drop()
     def cleanup(self):
         pass
